Nim game
Nim is a popular game for beginner programmers to learn basics of any programming language. Implementation of this game is not efficient, however, for the first time programming it looks ok for me. At least all the rules of the game is conserved and one can copy and play it. One can easily google the rules and winning conditions of this game. This code was developed as part of the university task.